src/main/kotlin/com/github/caay2000/ttk/context/entity/application/update/EntityUpdateUnloaderService.kt
package com.github.caay2000.ttk.context.entity.application.update

import arrow.core.Either
import com.github.caay2000.ttk.context.entity.domain.Entity
import com.github.caay2000.ttk.context.entity.domain.EntityException
import com.github.caay2000.ttk.context.entity.domain.EntityUpdateUnloaderServiceEntityException

class EntityUpdateUnloaderService {

    fun invoke(initialEntity: Entity): Either<EntityException, Entity> =
        Either.catch { initialEntity.updateUnload() }
            .mapLeft { EntityUpdateUnloaderServiceEntityException(it) }
}
